Upload tab cai dat

+++ DiemDanhChamCong/Models/Loaica.cs
@@ -13,7 +13,7 @@ public partial class Loaica
 
     public TimeOnly Ra { get; set; }
 
-    public long? HeSo { get; set; }
+    public double? HeSo { get; set; }
 
     public virtual ICollection<Chamcong> Chamcongs { get; set; } = new List<Chamcong>();
 }
