+++ DiemDanhChamCong/Models/Loaicong.cs
@@ -11,7 +11,7 @@ public partial class Loaicong
 
     public DateOnly Ngay { get; set; }
 
-    public long? HeSo { get; set; }
+    public double? HeSo { get; set; }
 
     public virtual ICollection<Chamcong> Chamcongs { get; set; } = new List<Chamcong>();
 }
